// frontend/src/api/models/manualBrokeragePortfolioResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { Account } from './account'
import type { InvestmentHoldingsResponse } from './investmentHoldingsResponse'

export interface ManualBrokeragePortfolioResponse {
  account: Account
  snapshot: InvestmentHoldingsResponse
  staleSymbols: string[]
}